Break down ModuleOverviewWidget into smaller components.
